# apps/tickets/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/tickets": {
      "post": {
        "operationId": "TicketsController_create",
        "summary": "Create a ticket - Scope : tickets:create_one",
        "description": "Request creation of a ticket",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTicketDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Ticket created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketDto"
                }
              }
            }
          }
        },
        "tags": ["tickets"],
        "security": [
          {
            "session": []
          },
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "TicketsController_find",
        "summary": "Find tickets - Scope : tickets:read_many",
        "description": "Filter tickets",
        "parameters": [
          {
            "name": "start_key",
            "required": false,
            "in": "query",
            "style": "deepObject",
            "explode": true,
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/StartKeyDto"
              }
            }
          },
          {
            "name": "skip",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "style": "deepObject",
            "explode": false,
            "schema": {
              "$ref": "#/components/schemas/SortDto"
            }
          },
          {
            "name": "filter",
            "required": false,
            "in": "query",
            "style": "deepObject",
            "explode": true,
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/FilterDto"
              }
            }
          },
          {
            "name": "projection",
            "required": false,
            "in": "query",
            "style": "deepObject",
            "explode": true,
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/ProjectionDto"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tickets found",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/PaginatedDto"
                    },
                    {
                      "properties": {
                        "results": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/TicketDto"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": ["tickets"]
      }
    },
    "/api/tickets/{id}": {
      "get": {
        "operationId": "TicketsController_findById",
        "summary": "Find a ticket - Scope : tickets:read_one",
        "description": "Request a ticket by id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ticket found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketDto"
                }
              }
            }
          }
        },
        "tags": ["tickets"]
      },
      "patch": {
        "operationId": "TicketsController_updateById",
        "summary": "Update a ticket - Scope : tickets:update_one",
        "description": "Update a ticket by id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTicketDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ticket updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketDto"
                }
              }
            }
          }
        },
        "tags": ["tickets"],
        "security": [
          {
            "session": []
          },
          {
            "bearer": []
          }
        ]
      }
    }
  },
  "info": {
    "title": "Tickets API",
    "description": "Ticketing tickets API description",
    "version": "0.2.0",
    "contact": {}
  },
  "tags": [
    {
      "name": "tickets",
      "description": ""
    }
  ],
  "servers": [
    {
      "url": "http://localhost:3010"
    },
    {
      "url": "http://localhost:8080"
    },
    {
      "url": "http://ticketing.dev"
    }
  ],
  "components": {
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session"
      },
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "NextPaginationDto": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string"
          },
          "value": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "number"
              }
            ]
          }
        },
        "required": ["key", "value"]
      },
      "PaginatedDto": {
        "type": "object",
        "properties": {
          "next": {
            "minItems": 1,
            "maxItems": 2,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NextPaginationDto"
            }
          }
        },
        "required": ["next"]
      },
      "CreateTicketDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Ticket title",
            "minLength": 3,
            "maxLength": 56
          },
          "price": {
            "type": "number",
            "description": "Ticket price",
            "minimum": 0
          }
        },
        "required": ["title", "price"]
      },
      "TicketDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Ticket database identifier"
          },
          "title": {
            "type": "string",
            "description": "Ticket title",
            "minLength": 3,
            "maxLength": 56
          },
          "price": {
            "type": "number",
            "description": "Ticket price",
            "minimum": 0
          },
          "version": {
            "type": "number",
            "description": "Ticket version represented by a number incremented at each updated"
          },
          "userId": {
            "type": "string",
            "description": "Ticket creator id"
          },
          "orderId": {
            "type": "string",
            "description": "Ticket reservation order id"
          }
        },
        "required": ["id", "title", "price", "version", "userId"]
      },
      "StartKeyDto": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string"
          },
          "value": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "number"
              }
            ]
          }
        },
        "required": ["key", "value"]
      },
      "SortDto": {
        "type": "object",
        "properties": {
          "field": {
            "type": "string",
            "minLength": 2,
            "maxLength": 20
          },
          "order": {
            "type": "number"
          }
        },
        "required": ["field", "order"]
      },
      "FilterOperatorType": {
        "type": "string",
        "enum": ["eq", "gt", "gte", "lt", "lte", "regex", "in", "nin"]
      },
      "FilterDto": {
        "type": "object",
        "properties": {
          "name": {
            "oneOf": [
              {
                "type": "array",
                "items": {
                  "type": "string",
                  "minLength": 2,
                  "maxLength": 20
                }
              },
              {
                "type": "string",
                "minLength": 2,
                "maxLength": 20
              }
            ]
          },
          "value": {
            "type": "string",
            "minLength": 1,
            "maxLength": 20
          },
          "arr_value": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "number"
              }
            ]
          },
          "operator": {
            "default": "eq",
            "$ref": "#/components/schemas/FilterOperatorType"
          },
          "mode": {
            "oneOf": [
              {
                "enum": ["eq", "gt", "gte", "lt", "lte", "regex", "in", "nin"],
                "default": "eq"
              },
              {
                "type": "string"
              }
            ]
          }
        },
        "required": ["name", "value", "operator"]
      },
      "ProjectionDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 20
          },
          "mode": {
            "type": "number",
            "minimum": 0,
            "maximum": 1
          }
        },
        "required": ["name", "mode"]
      },
      "UpdateTicketDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Ticket title",
            "minLength": 3,
            "maxLength": 56
          },
          "price": {
            "type": "number",
            "description": "Ticket price",
            "minimum": 0
          },
          "orderId": {
            "type": "string",
            "description": "Ticket reservation order id"
          }
        }
      }
    }
  },
  "security": [
    {
      "session": []
    },
    {
      "bearer": []
    }
  ]
}